Добовляет модули в проект
